small update about std lib usage

+++ Rocket/Engine/Memory/Handle.h
@@ -1,6 +1,5 @@
 #pragma once
 #include "Core/Core.h"
-#include "Core/Template.h"
 
 namespace Rocket {
     // Use Handle for Memory Manage and hide real pointer
@@ -8,6 +7,6 @@ namespace Rocket {
         uint32_t index;     // index for memory pool
         uint16_t category;  // hash of category name
         uint16_t type;      // hash of type name
-        uint8_t pattern;    // hash of content
+        uint16_t pattern;    // hash of content
     } Handle;
 }
